feat: made form section
FormSection is almost entirely copied from the TeacherClassSection of the  Home page with some changes like not using the TeacherClassSection since it has a button that doesn't exist on this page. (the massive chunk of code at the top is the copied TeacherClassSection without the button) Since the data used in the TeacherClassSection of the homepage is exactly the same one we want to display here it should be all good.